requirements.txt
python-dotenv
asttokens==3.0.0
backcall==0.2.0
boto3==1.37.38
botocore==1.37.38
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
findspark==2.0.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipython-sql==0.5.0
jedi==0.19.2
jmespath==1.0.1
jupyter-client==8.6.3
jupyter-core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prettytable==3.11.0
prompt-toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
pygments==2.19.2
pyspark==3.5.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
s3transfer==0.11.5
six==1.17.0
sqlalchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
zipp==3.20.2